Introduce and require "admin access" permission for the admin area
The permission is checked both on login and on access with an existing
user session.

It is meant to be assigned to (and withdrawn from) users as part of the
"orga" role.

This change decouples admin access from the orga flag which is set on a
per-party basis.

As a result, a newly installed BYCEPS system no longer requires the
initial admin user to obtain any orga flag, which in turn required any
brand to be created (on the BYCEPS shell or via script) before the first
admin login became possible.