Drop the state enums from seat group models
A seat group is either available or occupied, nothing else. (This will
make more sense with the upcoming changes.)

If a means of reservation is to be implemented add a later point, it
probably should be done with an external entity that is considered by
business logic rather than by re-introducing a tri-state field.