Include current terms version in user creation form and submit it to the server
This prevents a new user from reading a terms version and accepting another, newer one that is introduced and supersedes the first one during the registration step.

As a result, a user only accepts the version that is presented for him/her to read, not the one that is marked as the current one at the time of user account creation form submission.