Don't (attempt to) highlight current party in list of attended parties
The current party can only appear in that list after it's over, but then
there is not so much of a point to emphasize it.

Apart from that, the check was broken anyway as it didn't compare by ID
and two different classes were involved (party entity vs. party tuple).