Trigger order actions in shop service, not via signals from blueprint
With regard to actions to be triggered on order cancelation, moving the
call site to the order service gives access to the specific resulting
payment state (canceled *before* or *after* being paid) which is not
available in the blueprint view, but essential for further processing.

Pass order tuple instead of just the ID to the execution function to
avoid mutual import between order service and order action service.