Pass time of newsletter subscription change expression to the newsletter service
The time of expression by user action should be closely matched. Thus, the date should be determined in the frontend, and passed to the service as an argument to avoid any delays if the service will/cannot be called or won't start acting instantly, especially in an asynchronously communicating setup.