Don't explicitly unsubscribe from brand newsletter on user sign up, inline function
Initially, persisting the choice to not subscribe to the newsletter on
user sign up was considered useful.

However, to compile a list of subscribers to a newsletter, this is not
necessary, so drop that overhead.

This means that it cannot be determined from the subscription records
alone anymore if a user isn't subscribed to a newsletter because no
initial subscription was made or the subscription choice hasn't been
persisted due to some technical error. Logs and other mechanisms should
avoid, handle, or at least document the existance of such errors.