Dispatch to site-specific template context processors
To support multiple sites in a single application instance, this
introduces a template context processor that dispatches to a site's
specific one based on the site ID.

Notably, this removes the need for the site ID to be available on
application start (request context is sufficient).

Imported site-specific processors are cached on the application
instance.