Do not search for orders in ticketing check-in form
This is necessary for now as the shop reference is about to be moved
away from the party entity. A direct connection between a party and a
shop should no longer exist to gain flexibility.

As a workaround, search for the user that supposedly placed an order,
then go to the orders tab on the admin user view for that user.