Require a news item's slug to be unique
To be able to look up a news item by its slug when multiple news
channels will be possible per site, the unique constraint over (channel
ID, slug) is no longer sufficient.

The brand ID seems like an option, but it is not part of the news items
table (it is only implicit through the channel ID, but that's okay).

For the time being, constraint slugs to be globally unique. Including
the date should already help a lot, and if not, (manually, or automate
it yourself) prepending/appending a random token is also possible to
achieve uniqueness.

DDL to replace the slug index and multi-column constraint with a unique
index:

    ALTER TABLE news_items DROP CONSTRAINT news_items_channel_id_slug_key;
    DROP INDEX ix_news_items_slug;
    CREATE UNIQUE INDEX ix_news_items_slug ON news_items (slug);