Make database table creation script load all database models when called outside of project root
Previously, when the script was called from, say, the `scripts/` path,
the file-based approach to look for database model modules would search
just below the scripts path and thus would not find any matching models.

Now the search path is relative to the path of the `byceps` package, and
should find the modules even if the script is called from another path.