Remove creator from user avatar
It is no longer necessary as proper log events are now created for user
avatar updated (which include the creator).

This also removes the foreign key from the user avatar table to the user
table which would create a circular dependency with the upcoming change
that replaces the user/avatar cross table with a foreign key reference
from the user table to the user avatar table.

DDL:

    ALTER TABLE user_avatars DROP COLUMN creator_id;