Switch to association tables for featured news images
SQL to remove the superseded column (you should have filled the new
table from this column as detailed in the respective commit the other
day):

    ALTER TABLE news_items DROP COLUMN featured_image_id;

BTW, this is BYCEPS commit #10,000! \o/
It's been a while since the first commit in 2014, more than 9 years ago.
A very big thank you to all users and supporters!