Remove metrics sub-app hack, move endpoint to `/metrics/`
The hack using an empty string as path in the app dispatcher did not
work as expected. As soon as metrics were enabled, the admin app's
endpoint returned 404, making the admin UI inaccessible.

Due to checks in Werkzeug, the path of the metrics endpoint has changed
from `/metrics` to `/metrics/`.