Register order action procedures without payment state
Until now, order actions were defined in the database as a combination
of procedure name and payment state. If there was nothing to do on order
cancellation, there would be no procedure registered for target payment
state `canceled_after_paid`.

This changes order action definition so that in the database there is
only a procedure name without any payment state while the specific
function to call for a specific payment state is defined in code.

SQL/DDL:

    DELETE FROM shop_order_actions WHERE procedure IN ('revoke_ticket_bundles', 'revoke_tickets');

    ALTER TABLE shop_order_actions DROP COLUMN payment_state;